move microsd toast view to screen, using display to render components
