+++ src/seedsigner/gui/screens/screen.py
@@ -10,8 +10,7 @@ from seedsigner.models.threads import BaseThread
 from seedsigner.models.encode_qr import EncodeQR
 from seedsigner.models.settings import Settings, SettingsConstants
 
-from ..components import (FontAwesomeIconConstants, GUIConstants, BaseComponent, Button, Icon, IconButton, LargeIconButton, SeedSignerCustomIconConstants, TopNav,
-    TextArea, load_image)
+from ..components import (FontAwesomeIconConstants, GUIConstants, BaseComponent, Button, Icon, IconButton, LargeIconButton, SeedSignerCustomIconConstants, TopNav, TextArea, load_image, ToastOverlay)
 
 from seedsigner.hardware.buttons import HardwareButtonsConstants, HardwareButtons
 
@@ -1081,3 +1080,38 @@ class KeyboardScreen(BaseTopNavScreen):
                 self.title = f"Roll {self.cursor_position + 1}"
         """
         return False
+
+class MicroSDToastScreen(BaseScreen):
+    """
+        This screen is an overlay with special behavior with the ToastOverlay component. The ToastOverlay component overides all button
+        input and captures the existing screen content and stashes it to be restored once X second passes or any button is pressed. The
+        display method on this screen will not complete until after the ToastOverlay render method is complete it's takeover of the screen.
+    """
+    def __init__(self, action):
+        self.action = action
+        self.toast = None
+        super().__init__()
+    
+    def _run(self):
+        return
+    
+    def _render(self):
+        from seedsigner.hardware.microsd import MicroSD
+        
+        if self.action == MicroSD.ACTION__REMOVED:
+        
+            self.toast = ToastOverlay(
+                icon_name=FontAwesomeIconConstants.SDCARD,
+                color=GUIConstants.NOTIFICATION_COLOR,
+                label_text="MicroSD removed"
+            )
+        
+        elif self.action == MicroSD.ACTION__INSERTED:
+            
+            self.toast = ToastOverlay(
+                icon_name=FontAwesomeIconConstants.SDCARD,
+                color=GUIConstants.NOTIFICATION_COLOR,
+                label_text="MicroSD inserted"
+            )
+        
+        self.toast.render()
